Add searchoptions from rental screen to user screen. Move search functions to separate class.
